fixed absolute value calculator
