+++ src/js/map/choropleth/absolute_value_calculator.js
@@ -1,9 +1,11 @@
-import {percFmt, numFmt} from '../../utils'
+import {numFmt, filterAndSumGeoCounts} from '../../utils'
 
 export default class AbsoluteValueCalculator {
     constructor() {}
-    calculate(args) {
-        const result = Object.entries(args.data).map(childGeography => {
+    calculate(childData, primaryGroup, selectedSubindicator) {
+        let sumData = filterAndSumGeoCounts(childData, primaryGroup, selectedSubindicator);
+
+        const result = Object.entries(sumData).map(childGeography => {
             const code = childGeography[0];
             const count = childGeography[1];
             return {code: code, val: count};
